+++ Maven项目/Utunan/src/main/webapp/WEB-INF/view/share/download.jsp
@@ -212,7 +212,7 @@
         $.ajax({
             type: "POST",//方法类型
             dataType: "json",//预期服务器返回的数据类型
-            url: "/school/popsupLogin" ,//url
+            url: "/popsupLogin" ,//url
             data: $('#loginform').serialize(),
             success: function (result) {
                if (result==true){
