+++ Maven项目/Utunan/src/main/webapp/WEB-INF/view/share/right.jsp
@@ -59,7 +59,7 @@
             //几个参数需要注意一下
             type: "POST",//方法类型
             dataType: "json",//预期服务器返回的数据类型
-            url: "/school/popsupLogin" ,//url
+            url: "/popsupLogin" ,//url
             data: $('#loginform').serialize(),
             success: function (result) {
                 console.log(result);//打印服务端返回的数据(调试用)
